Added the OR, but not quite working
It works, but the structure of a surrounding group is required.
I need to figure out a way to delimit the "parsing context", so
within a group the "or stack" is processed.

The way the `or stack` is processed is also not super elegant either.
I wish I could "sub-group" parse left-to-right for the ors, but the
current structure does not allow that.

Maybe I'm too tired to figure it out right now.